+++ servers/mcp-neo4j-data-modeling/CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## v0.1.0
 
-* Basic functionality - Expose schemas for Data Model, Node, Relationship and Property
+* Basic functionality 
+  * Expose schemas for Data Model, Node, Relationship and Property
+  * Validation tools
 * Visualize data model in interactive browser window   
 * Import / Export from Arrows web application